Auto Fixed
Found what was wrong with turning the wrong way and fixed it. Pretty simple change thankfully, I was being dumb when I made it ig. Also the slide position transfer works now so no more (or at least a lot less) manual resets thankfully.
TODO: Will want to transfer slide position in this way in DeliveryTeleop IF GridTeleOp doesn't wind up working out. Juuuust in case!

+++ PoseTransfer.java
@@ -8,6 +8,6 @@ public class PoseTransfer {
     public static int idealGridCoordX = 0;
     public static int idealGridCoordY = 0;
     public static int idealGridAngle = 0;
-    public static int slidePos = 0;
+    public static double slidePos = 0;
 
 }
